+++ packages/components-native/src/__mocks__/__mocks.ts
@@ -1,10 +1,14 @@
 import { MissingTranslationError } from "react-intl";
 import mockRNLocalize from "react-native-localize/mock";
+import * as ReactNative from "react-native";
+import { MockModal } from "./MockModal";
 
 jest.mock("react-native-localize", () => ({
   ...mockRNLocalize,
   getTimeZone: () => "UTC",
 }));
+
+jest.mock("react-native/Libraries/Modal/Modal", () => MockModal);
 
 jest.mock("react-native-reanimated", () => {
   const reanimated = require("react-native-reanimated/mock");
@@ -44,3 +48,15 @@ jest.mock("react-intl", () => {
     useIntl: () => intl,
   };
 });
+
+jest.mock("../hooks/useIsScreenReaderEnabled", () => ({
+  useIsScreenReaderEnabled: () => false,
+}));
+
+// By default React Native mocks the return value of `addEventListener` using
+// jest.fn(). See https://github.com/facebook/react-native/blob/0.69-stable/jest/setup.js#L124
+// This provides an implementation of .remove for useEffect's that do subscribe and
+// call .remove on tear down.
+jest.spyOn(ReactNative.AccessibilityInfo, "addEventListener").mockReturnValue({
+  remove: jest.fn(),
+} as unknown as ReactNative.EmitterSubscription);